+++ Network/API/Telegram/Bot/Object/Update/Message/Content/File/Audio.hs
@@ -9,12 +9,14 @@ import "base" Data.Functor ((<$>))
 import "base" Text.Show (Show)
 import "text" Data.Text (Text)
 
+import Network.API.Telegram.Bot.Object.Update.Message.Content.File.URI (URI)
 import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Duration (Duration)
 
-data Audio = Audio Duration (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Int)
+data Audio = Audio URI Duration (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Int)
 	deriving Show
 
 instance FromJSON Audio where
 	parseJSON = withObject "Audio" $ \v -> Audio
-		<$> v .: "duration" <*> v .:? "performer" <*> v .:? "title"
+		<$> v .: "file_id" <*> v .: "duration"
+		<*> v .:? "performer" <*> v .:? "title"
 		<*> v .:? "mime_type" <*> v .:? "file_size"