+++ Network/API/Telegram/Bot/Object/Update/Message/Content/File/Voice.hs
@@ -9,10 +9,11 @@ import "base" Data.Functor ((<$>))
 import "base" Text.Show (Show)
 import "text" Data.Text (Text)
 
+import Network.API.Telegram.Bot.Object.Update.Message.Content.File.URI (URI)
 import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Duration (Duration)
 
-data Voice = Voice Duration (Maybe Text) (Maybe Int) deriving Show
+data Voice = Voice URI Duration (Maybe Text) (Maybe Int) deriving Show
 
 instance FromJSON Voice where
-	parseJSON = withObject "Voice" $ \v -> Voice <$>
-		v .: "duration" <*> v .:? "mime_type" <*> v .:? "file_size"
+	parseJSON = withObject "Voice" $ \v -> Voice <$> v .: "file_id"
+		<*> v .: "duration" <*> v .:? "mime_type" <*> v .:? "file_size"